server/legacy_ops_environment.py
import json
import posixpath
import base64
from uuid import uuid4

from openenv.core.env_server.interfaces import Environment
from openenv.core.env_server.types import State

try:
    from ..models import LegacyOpsAction, LegacyOpsObservation
except ImportError:
    from models import LegacyOpsAction, LegacyOpsObservation

class LegacyOpsEnvironment(Environment):
    SUPPORTS_CONCURRENT_SESSIONS: bool = True

    def __init__(self):
        self._state = State(episode_id=str(uuid4()), step_count=0)
        
        # Safely try to load config, but don't crash if validator doesn't mount it
        try:
            with open("assets/campaign_config.json", "r") as f:
                self.config = json.load(f)
        except Exception:
            self.config = {}
            
        self.filesystem = self.config.get("filesystem", {})
        self.global_hint = self.config.get("global_hint", "System Breach Detected.")
        
        self.expected_flags = [
            "FLAG{fragmented_auth_bypassed}",
            "FLAG{multi_layer_crypto_cracked}",
            "FLAG{root_environment_secured}",
            "FLAG{integrity_recovered}",
            "FLAG{access_control_restored}",
            "FLAG{threat_neutralized}"
        ]
        self.setup_game_state()

    def setup_game_state(self):
        self.cwd = "/"
        self.stdout = f"--- MISSION START ---\n{self.global_hint}"
        self.stderr = ""
        self.current_phase = 0
        self.total_reward = 0.0
        self.game_done = False
        self.nginx_restored = False
        self.shadow_secured = False
        self.malware_removed = False
        self.action_history = set()

    def reset(self) -> LegacyOpsObservation:
        self._state = State(episode_id=str(uuid4()), step_count=0)
        self.setup_game_state()
        
        return LegacyOpsObservation(
            cwd=self.cwd,
            stdout=self.stdout,
            stderr=self.stderr,
            current_phase=self.current_phase,
            done=self.game_done,
            reward=0.0
        )

    def _get_fs_node(self, path):
        if path in ["/", ""]: return self.filesystem
        parts = [p for p in path.strip("/").split("/") if p]
        curr = self.filesystem
        for p in parts:
            if isinstance(curr, dict) and p in curr: curr = curr[p]
            else: return None
        return curr

    def _get_file_content(self, node):
        if isinstance(node, dict) and "content" in node:
            metadata = node.get("metadata", {})
            if metadata.get("required_phase", 0) > self.current_phase:
                return None, f"ACCESS DENIED: Phase {metadata['required_phase']} required."
            return str(node["content"]), ""
        elif isinstance(node, str): return node, ""
        return None, "Path is a directory or does not exist."

    def step(self, action: LegacyOpsAction) -> LegacyOpsObservation:
        self._state.step_count += 1
        
        if self.game_done:
            return LegacyOpsObservation(
                cwd=self.cwd, stdout="", stderr="Mission Complete.", 
                current_phase=self.current_phase, done=True, reward=0.0
            )

        cmd = getattr(action, "command", "")
        target = getattr(action, "target", "")
        
        self.stdout, self.stderr = "", ""
        action_signature = f"{self.current_phase}|{self.cwd}|{cmd}|{target}"
        step_reward = -0.01  
        
        target_path = posixpath.normpath(posixpath.join(self.cwd, target or "")).lstrip('/')

        if cmd == "ls":
            node = self._get_fs_node(target_path)
            if isinstance(node, dict) and "content" not in node: 
                self.stdout = "\n".join([k for k in node.keys() if k != "metadata"])
            else: self.stderr = f"ls: {target}: No such directory"
                
        elif cmd == "cd":
            node = self._get_fs_node(target_path)
            if node is not None and isinstance(node, dict) and "content" not in node:
                self.cwd = "/" + target_path
            else: self.stderr = f"cd: {target}: Directory not found"

        elif cmd == "cat":
            node = self._get_fs_node(target_path)
            content, err = self._get_file_content(node)
            if content: self.stdout = content
            else: self.stderr = f"cat: {target}: {err}"

        elif cmd == "env":
            if self.current_phase < 2: self.stderr = "SYSTEM ERROR: Env monitoring offline."
            else: self.stdout = "USER=agent\nSHELL=/bin/bash\nROOT_PIN=FLAG{root_environment_secured}\nPATH=/usr/bin\nSESSION=active"

        elif cmd == "decode":
            if self.current_phase < 1: self.stderr = "SYSTEM ERROR: Utility offline."
            else:
                try: self.stdout = base64.b64decode(target).decode('utf-8')
                except: self.stderr = "decode: Invalid Base64 data."
                    
        elif cmd == "cp":
            if "nginx_clean.conf" in target_path and "nginx.conf" in target_path:
                self.nginx_restored = True
                self.stdout = "STATE UPDATE: nginx.conf overwritten. Validation: FLAG{integrity_recovered}"
            else: self.stderr = f"cp: invalid target '{target}'"

        elif cmd == "chmod":
            if "600" in (target or "") and "shadow" in target_path:
                self.shadow_secured = True
                self.stdout = "STATE UPDATE: /etc/shadow secured. Validation: FLAG{access_control_restored}"
            else: self.stderr = "chmod: Invalid permissions or target."

        elif cmd == "rm":
            if "safe_cache" in target_path: self.stderr = "WARNING: Cannot delete safe system file!"
            elif "sysupdater" in target_path:
                self.malware_removed = True
                self.stdout = "STATE UPDATE: Malware removed. Validation: FLAG{threat_neutralized}"
            else: self.stderr = f"rm: cannot remove '{target}'"

        elif cmd == "submit_flag":
            try:
                expected_flag = self.expected_flags[self.current_phase]
                if self.current_phase == 3 and not self.nginx_restored:
                    self.stderr = "VALIDATION FAILED: nginx.conf state still corrupted."
                    step_reward = -0.05
                elif self.current_phase == 4 and not self.shadow_secured:
                    self.stderr = "VALIDATION FAILED: /etc/shadow vulnerable."
                    step_reward = -0.05
                elif self.current_phase == 5 and not self.malware_removed:
                    self.stderr = "VALIDATION FAILED: Malware 'sysupdater' still active."
                    step_reward = -0.05
                elif target == expected_flag:
                    step_reward = 0.99  
                    self.current_phase += 1
                    self.stdout = f"[SUCCESS] Step {self.current_phase}/6 complete."
                    if self.current_phase >= 6:
                        self.stdout += " 🏆 MISSION SECURED."
                        self.game_done = True
                else:
                    self.stderr = "SUBMISSION FAILED: Invalid flag."
                    step_reward = -0.05
            except IndexError:
                self.stderr = "ERROR: All phases complete."
        else:
            if cmd not in ["grep", "hex_decode"]:
                self.stderr = f"bash: {cmd}: command not found"

        if action_signature in self.action_history:
            step_reward = 0.0
        else:
            self.action_history.add(action_signature)

        self.total_reward += step_reward

        return LegacyOpsObservation(
            cwd=self.cwd,
            stdout=self.stdout,
            stderr=self.stderr,
            current_phase=self.current_phase,
            done=self.game_done,
            reward=step_reward
        )

    @property
    def state(self) -> State:
        return self._state